+++ packages/api/src/main/messaging/messaging-connection.ts
@@ -6,13 +6,18 @@
  */
 import amqp from 'amqplib'
 
-const CONNECT_RETRY_INTERVAL = 5000
+import { Logging } from '../logging/logging.js'
+import { CONNECT_RETRY_INTERVAL } from './constants.js'
+
+const WAIT_FOR_CONFIRMS_TIMEOUT = 3000
 
 interface MessagingConnectionConfig {
-  callback?: () => void | Promise<void>
+  onChannelReady?: (channel: amqp.ConfirmChannel) => void | Promise<void>
   url: string
   retry: boolean
-  socketOptions: any
+  socketOptions: {
+    [key: string]: any
+  }
 }
 
 /**
@@ -20,16 +25,39 @@ interface MessagingConnectionConfig {
  */
 class MessagingConnection {
   private readonly config: MessagingConnectionConfig
-  private actualConnection?: amqp.Connection
-  private actualChannel?: amqp.ConfirmChannel
+  private readonly logging: Logging
+  private connection?: amqp.Connection
+  private _channel?: amqp.ConfirmChannel
   private channelPromise?: Promise<amqp.ConfirmChannel>
 
   public constructor(config: MessagingConnectionConfig) {
     this.config = config
+    this.logging = new Logging({
+      component: 'messaging-connection',
+      isRemoteLoggingEnabled: false
+    })
+  }
+
+  private setUpErrorHandlers() {
+    this.connection?.once('error', (err) => {
+      this.logging.error('Messaging connection emitted an error')
+      this.logging.error(err)
+
+      this.close()
+    })
+
+    this._channel?.once('error', (err) => {
+      this.logging.error('Messaging channel emitted an error')
+      this.logging.error(err)
+
+      this.close()
+    })
   }
 
   /**
-   * TODO: docs
+   * The underlying channel of the messaging connection. This lazily establishes the connection to
+   * the message broker. Once the connection is established, the callback provided in the config
+   * object to the constructor will be invoked prior to the channel being returned to the caller.
    *
    * **NOTE:** This can throw if there is an issue either connecting or running the provided
    * callback.
@@ -42,17 +70,31 @@ class MessagingConnection {
 
     // No existing connection. Set up a promise of a new one
     this.channelPromise = (async () => {
-      while (!this.actualConnection || !this.actualChannel) {
-        try {
-          this.actualConnection = await amqp.connect(
-            this.config.url,
-            this.config.socketOptions ?? {}
-          )
-          this.actualChannel = await this.actualConnection.createConfirmChannel()
-        } catch (e) {
-          console.error('Could not connect to messaging service', e)
+      let isFirstTry = true
 
-          await this.close()
+      while (!this.connection || !this._channel) {
+        try {
+          if (isFirstTry) {
+            // Use a URL object to avoid logging usernames and passwords included in the url
+            const url = new URL(this.config.url.replace(/^amqp/, 'http'))
+            this.logging.info('Connecting to message broker at ' + url.host)
+          }
+
+          this.connection = await amqp.connect(this.config.url, this.config.socketOptions)
+
+          this.setUpErrorHandlers()
+
+          this._channel = await this.connection.createConfirmChannel()
+        } catch (e) {
+          if (isFirstTry) {
+            this.logging.error('Could not connect to messaging service')
+            if (e instanceof Error) {
+              this.logging.error(e)
+            }
+          }
+
+          // Cleanup, but don't destory the promise we're currently inside
+          await this.close(isFirstTry, false)
 
           if (this.config.retry) {
             // Retry again after a few seconds
@@ -63,30 +105,76 @@ class MessagingConnection {
             throw e
           }
         }
+
+        isFirstTry = false
       }
 
-      this.config.callback?.bind(this)()
+      this.logging.info('Successfully connected to message broker')
 
-      return this.actualChannel
+      if (this.config.onChannelReady) {
+        this.logging.debug(`-> callback: ${this.config.onChannelReady.name}()`)
+        this.config.onChannelReady(this._channel)
+        this.logging.debug(`<- callback: ${this.config.onChannelReady.name}()`)
+      }
+
+      return this._channel
     })()
 
     return this.channelPromise
   }
 
-  public async close() {
-    console.warn('Closing messaging connection')
+  public async close(logStatus = true, destroyPromise = true) {
+    logStatus && this.logging.debug('Closing messaging connection')
 
     try {
-      await this.actualChannel?.close()
-      await this.actualConnection?.close()
+      await Promise.race([
+        this._channel?.waitForConfirms(),
+        new Promise((_resolve, reject) => {
+          setTimeout(
+            () => reject(new Error('Timed out waiting for confirms')),
+            WAIT_FOR_CONFIRMS_TIMEOUT
+          )
+        })
+      ])
     } catch (e) {
-      console.warn('Exception caught while trying to close an existing messaging connection', e)
+      if (logStatus) {
+        this.logging.warn(
+          'Exception caught waiting for confirms while trying to close messaging connection'
+        )
+        if (e instanceof Error) {
+          this.logging.warn(e)
+        }
+      }
+    }
+
+    try {
+      await this._channel?.close()
+    } catch (e) {
+      if (logStatus) {
+        this.logging.warn(
+          'Exception caught closing channel while trying to close messaging connection'
+        )
+        if (e instanceof Error) {
+          this.logging.warn(e)
+        }
+      }
+    }
+
+    try {
+      await this.connection?.close()
+    } catch (e) {
+      if (logStatus) {
+        this.logging.warn('Exception caught while trying to close messaging connection')
+        if (e instanceof Error) {
+          this.logging.warn(e)
+        }
+      }
     }
 
     // Future uses of this connection will trigger a reconnect
-    delete this.actualChannel
-    delete this.actualConnection
-    delete this.channelPromise
+    delete this._channel
+    delete this.connection
+    destroyPromise && delete this.channelPromise
   }
 }